Azure IA/img_azure/python/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-core==1.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
h11==0.14.0
idna==3.6
isodate==0.6.1
pydantic==2.6.3
pydantic_core==2.16.3
python-multipart==0.0.9
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
wheel==0.41.2
